backend/requirements.txt
fastapi
uvicorn
httpx
google-generativeai
pydantic
